Move benchmark specific code from LocalQueryRunner to benchmarks
