+++ presto-benchmark/src/main/java/com/facebook/presto/benchmark/AbstractOperatorBenchmark.java
@@ -14,31 +14,67 @@
 package com.facebook.presto.benchmark;
 
 import com.facebook.presto.Session;
+import com.facebook.presto.execution.Lifespan;
 import com.facebook.presto.execution.TaskId;
 import com.facebook.presto.execution.TaskStateMachine;
 import com.facebook.presto.memory.DefaultQueryContext;
 import com.facebook.presto.memory.MemoryPool;
+import com.facebook.presto.metadata.Metadata;
+import com.facebook.presto.metadata.QualifiedObjectName;
+import com.facebook.presto.metadata.Split;
+import com.facebook.presto.metadata.TableHandle;
+import com.facebook.presto.metadata.TableLayoutHandle;
+import com.facebook.presto.metadata.TableLayoutResult;
 import com.facebook.presto.operator.Driver;
+import com.facebook.presto.operator.DriverContext;
+import com.facebook.presto.operator.FilterAndProjectOperator;
+import com.facebook.presto.operator.Operator;
+import com.facebook.presto.operator.OperatorContext;
 import com.facebook.presto.operator.OperatorFactory;
+import com.facebook.presto.operator.PageSourceOperator;
 import com.facebook.presto.operator.TaskContext;
 import com.facebook.presto.operator.TaskStats;
+import com.facebook.presto.operator.project.InterpretedPageProjection;
+import com.facebook.presto.operator.project.PageProcessor;
+import com.facebook.presto.operator.project.PageProjection;
 import com.facebook.presto.security.AllowAllAccessControl;
+import com.facebook.presto.spi.ColumnHandle;
+import com.facebook.presto.spi.ColumnMetadata;
+import com.facebook.presto.spi.ConnectorPageSource;
+import com.facebook.presto.spi.Constraint;
 import com.facebook.presto.spi.QueryId;
 import com.facebook.presto.spi.memory.MemoryPoolId;
 import com.facebook.presto.spi.type.Type;
 import com.facebook.presto.spiller.SpillSpaceTracker;
+import com.facebook.presto.split.SplitSource;
+import com.facebook.presto.sql.planner.Symbol;
+import com.facebook.presto.sql.planner.optimizations.HashGenerationOptimizer;
 import com.facebook.presto.sql.planner.plan.PlanNodeId;
+import com.facebook.presto.sql.tree.Expression;
+import com.facebook.presto.sql.tree.SymbolReference;
 import com.facebook.presto.testing.LocalQueryRunner;
 import com.facebook.presto.transaction.TransactionId;
+import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
+import com.google.common.collect.Iterables;
 import io.airlift.stats.CpuTimer;
 import io.airlift.stats.TestingGcMonitor;
 import io.airlift.units.DataSize;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 
+import static com.facebook.presto.SystemSessionProperties.getFilterAndProjectMinOutputPageRowCount;
+import static com.facebook.presto.SystemSessionProperties.getFilterAndProjectMinOutputPageSize;
+import static com.facebook.presto.spi.connector.ConnectorSplitManager.SplitSchedulingStrategy.UNGROUPED_SCHEDULING;
+import static com.facebook.presto.spi.connector.NotPartitionedPartitionHandle.NOT_PARTITIONED;
+import static com.facebook.presto.spi.type.BigintType.BIGINT;
 import static com.facebook.presto.testing.TestingSession.testSessionBuilder;
+import static com.google.common.base.Preconditions.checkArgument;
+import static com.google.common.base.Verify.verify;
+import static io.airlift.concurrent.MoreFutures.getFutureValue;
 import static io.airlift.stats.CpuTimer.CpuDuration;
 import static io.airlift.units.DataSize.Unit.BYTE;
 import static io.airlift.units.DataSize.Unit.GIGABYTE;
@@ -89,14 +125,115 @@ public abstract class AbstractOperatorBenchmark
         super.tearDown();
     }
 
-    protected OperatorFactory createTableScanOperator(int operatorId, PlanNodeId planNodeId, String tableName, String... columnNames)
+    protected final OperatorFactory createTableScanOperator(int operatorId, PlanNodeId planNodeId, String tableName, String... columnNames)
     {
-        return localQueryRunner.createTableScanOperator(session, operatorId, planNodeId, tableName, columnNames);
+        checkArgument(session.getCatalog().isPresent(), "catalog not set");
+        checkArgument(session.getSchema().isPresent(), "schema not set");
+
+        // look up the table
+        Metadata metadata = localQueryRunner.getMetadata();
+        QualifiedObjectName qualifiedTableName = new QualifiedObjectName(session.getCatalog().get(), session.getSchema().get(), tableName);
+        TableHandle tableHandle = metadata.getTableHandle(session, qualifiedTableName).orElse(null);
+        checkArgument(tableHandle != null, "Table %s does not exist", qualifiedTableName);
+
+        // lookup the columns
+        Map<String, ColumnHandle> allColumnHandles = metadata.getColumnHandles(session, tableHandle);
+        ImmutableList.Builder<ColumnHandle> columnHandlesBuilder = ImmutableList.builder();
+        ImmutableList.Builder<Type> columnTypesBuilder = ImmutableList.builder();
+        for (String columnName : columnNames) {
+            ColumnHandle columnHandle = allColumnHandles.get(columnName);
+            checkArgument(columnHandle != null, "Table %s does not have a column %s", tableName, columnName);
+            columnHandlesBuilder.add(columnHandle);
+            ColumnMetadata columnMetadata = metadata.getColumnMetadata(session, tableHandle, columnHandle);
+            columnTypesBuilder.add(columnMetadata.getType());
+        }
+        List<ColumnHandle> columnHandles = columnHandlesBuilder.build();
+        List<Type> columnTypes = columnTypesBuilder.build();
+
+        // get the split for this table
+        List<TableLayoutResult> layouts = metadata.getLayouts(session, tableHandle, Constraint.alwaysTrue(), Optional.empty());
+        Split split = getLocalQuerySplit(session, layouts.get(0).getLayout().getHandle());
+
+        return new OperatorFactory()
+        {
+            @Override
+            public List<Type> getTypes()
+            {
+                return columnTypes;
+            }
+
+            @Override
+            public Operator createOperator(DriverContext driverContext)
+            {
+                OperatorContext operatorContext = driverContext.addOperatorContext(operatorId, planNodeId, "BenchmarkSource");
+                ConnectorPageSource pageSource = localQueryRunner.getPageSourceManager().createPageSource(session, split, columnHandles);
+                return new PageSourceOperator(pageSource, operatorContext);
+            }
+
+            @Override
+            public void noMoreOperators()
+            {
+            }
+
+            @Override
+            public OperatorFactory duplicate()
+            {
+                throw new UnsupportedOperationException();
+            }
+        };
     }
 
-    protected OperatorFactory createHashProjectOperator(int operatorId, PlanNodeId planNodeId, List<Type> types)
+    private Split getLocalQuerySplit(Session session, TableLayoutHandle handle)
     {
-        return localQueryRunner.createHashProjectOperator(session, operatorId, planNodeId, types);
+        SplitSource splitSource = localQueryRunner.getSplitManager().getSplits(session, handle, UNGROUPED_SCHEDULING);
+        List<Split> splits = new ArrayList<>();
+        while (!splitSource.isFinished()) {
+            splits.addAll(getNextBatch(splitSource));
+        }
+        checkArgument(splits.size() == 1, "Expected only one split for a local query, but got %s splits", splits.size());
+        return splits.get(0);
+    }
+
+    private static List<Split> getNextBatch(SplitSource splitSource)
+    {
+        return getFutureValue(splitSource.getNextBatch(NOT_PARTITIONED, Lifespan.taskWide(), 1000)).getSplits();
+    }
+
+    protected final OperatorFactory createHashProjectOperator(int operatorId, PlanNodeId planNodeId, List<Type> types)
+    {
+        ImmutableMap.Builder<Symbol, Type> symbolTypes = ImmutableMap.builder();
+        ImmutableMap.Builder<Symbol, Integer> symbolToInputMapping = ImmutableMap.builder();
+        ImmutableList.Builder<PageProjection> projections = ImmutableList.builder();
+        for (int channel = 0; channel < types.size(); channel++) {
+            Symbol symbol = new Symbol("h" + channel);
+            symbolTypes.put(symbol, types.get(channel));
+            symbolToInputMapping.put(symbol, channel);
+            projections.add(new InterpretedPageProjection(
+                    new SymbolReference(symbol.getName()),
+                    ImmutableMap.of(symbol, types.get(channel)),
+                    ImmutableMap.of(symbol, channel),
+                    localQueryRunner.getMetadata(),
+                    localQueryRunner.getSqlParser(),
+                    session));
+        }
+
+        Optional<Expression> hashExpression = HashGenerationOptimizer.getHashExpression(ImmutableList.copyOf(symbolTypes.build().keySet()));
+        verify(hashExpression.isPresent());
+        projections.add(new InterpretedPageProjection(
+                hashExpression.get(),
+                symbolTypes.build(),
+                symbolToInputMapping.build(),
+                localQueryRunner.getMetadata(),
+                localQueryRunner.getSqlParser(),
+                session));
+
+        return new FilterAndProjectOperator.FilterAndProjectOperatorFactory(
+                operatorId,
+                planNodeId,
+                () -> new PageProcessor(Optional.empty(), projections.build()),
+                ImmutableList.copyOf(Iterables.concat(types, ImmutableList.of(BIGINT))),
+                getFilterAndProjectMinOutputPageSize(session),
+                getFilterAndProjectMinOutputPageRowCount(session));
     }
 
     protected abstract List<Driver> createDrivers(TaskContext taskContext);
